Change all uint_fast32_t usage to use size_t instead
Because it's more concise, more descriptive, since most of the uses
are counts.

The "present" variable has been changed to bool instead.

+++ samples/pet/include/pet_types.h
@@ -31,7 +31,7 @@ extern "C" {
 
 struct Pet {
 	struct zcbor_string names[3];
-	uint_fast32_t names_count;
+	size_t names_count;
 	struct zcbor_string birthday;
 	enum {
 		_Pet_species_cat = 1,
